+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   - Add missing sources jar, required by JCenter
   - Change `KtsExec` `script` param type to `RegularFileProperty` (#7)
   - Skip task execution if no `script` param was provided (#8)
+  - `KtsExec` task depends only on host dependency (#9)
 
 ### [1.0.0] - 2019-03-24
   - Initial implementation